% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caminho.R
\name{caminho}
\alias{caminho}
\title{função para definir o diretório de cada tabela do datasus automaticamente}
\usage{
caminho(link, tabela)
}
\arguments{
\item{link}{é o link de download da tabela}

\item{tabela}{é o nome da tabela que está sendo baixada}
}
\value{
retorna o diretório que receberá o arquivo
}
\description{
função para definir o diretório de cada tabela do datasus automaticamente
}
